Improve comment detection.
Single line comments are not very useful as an indication of
documentation sections. They are often used to document a behaviour of
the function for improved readability of the source code, sometimes with
yaml syntax (lines containing e.g. a colon). To avoid confusing the
parser as to what is an internal comment and what is a styleguide
comment, we decided to keep styleguide comments in multiline comments and
internal comments in single line.